+++ crates/oxc_linter/src/rules/eslint/max_lines.rs
@@ -82,23 +82,29 @@ impl Rule for MaxLines {
     fn run_once(&self, ctx: &LintContext) {
         let comment_lines = if self.skip_comments {
             let mut comment_lines: usize = 0;
-            for (kind, span) in ctx.semantic().trivias().comments() {
-                if kind.is_single_line() {
-                    let comment_line =
-                        ctx.source_text()[..span.start as usize].lines().next_back().unwrap_or("");
+            for comment in ctx.semantic().trivias().comments() {
+                if comment.kind.is_single_line() {
+                    let comment_line = ctx.source_text()[..comment.span.start as usize]
+                        .lines()
+                        .next_back()
+                        .unwrap_or("");
                     if line_has_just_comment(comment_line, "//") {
                         comment_lines += 1;
                     }
                 } else {
-                    let mut start_line = ctx.source_text()[..span.start as usize].lines().count();
-                    let comment_start_line =
-                        ctx.source_text()[..span.start as usize].lines().next_back().unwrap_or("");
+                    let mut start_line =
+                        ctx.source_text()[..comment.span.start as usize].lines().count();
+                    let comment_start_line = ctx.source_text()[..comment.span.start as usize]
+                        .lines()
+                        .next_back()
+                        .unwrap_or("");
                     if !line_has_just_comment(comment_start_line, "/*") {
                         start_line += 1;
                     }
-                    let mut end_line = ctx.source_text()[..=span.end as usize].lines().count();
+                    let mut end_line =
+                        ctx.source_text()[..=comment.span.end as usize].lines().count();
                     let comment_end_line =
-                        ctx.source_text()[span.end as usize..].lines().next().unwrap_or("");
+                        ctx.source_text()[comment.span.end as usize..].lines().next().unwrap_or("");
                     if line_has_just_comment(comment_end_line, "*/") {
                         end_line += 1;
                     }