+++ crates/oxc_parser/src/js/statement.rs
@@ -32,8 +32,8 @@ impl<'a> ParserImpl<'a> {
         &mut self,
         is_top_level: bool,
     ) -> Result<(Vec<'a, Directive<'a>>, Vec<'a, Statement<'a>>)> {
-        let mut directives = self.ast.new_vec();
-        let mut statements = self.ast.new_vec();
+        let mut directives = self.ast.vec();
+        let mut statements = self.ast.vec();
 
         let mut expecting_directives = true;
         while !self.at(Kind::Eof) {
@@ -137,7 +137,7 @@ impl<'a> ParserImpl<'a> {
     pub(crate) fn parse_block(&mut self) -> Result<Box<'a, BlockStatement<'a>>> {
         let span = self.start_span();
         self.expect(Kind::LCurly)?;
-        let mut body = self.ast.new_vec();
+        let mut body = self.ast.vec();
         while !self.at(Kind::RCurly) && !self.at(Kind::Eof) {
             let stmt = self.parse_statement_list_item(StatementContext::StatementList)?;
             body.push(stmt);
@@ -457,7 +457,7 @@ impl<'a> ParserImpl<'a> {
             _ => return Err(self.unexpected()),
         };
         self.expect(Kind::Colon)?;
-        let mut consequent = self.ast.new_vec();
+        let mut consequent = self.ast.vec();
         while !matches!(self.cur_kind(), Kind::Case | Kind::Default | Kind::RCurly | Kind::Eof) {
             let stmt = self.parse_statement_list_item(StatementContext::StatementList)?;
             consequent.push(stmt);